+++ rust/sedona-spatial-join/src/physical_planner.rs
